Remove external interactors support in v2.
In lieu of a more fully-fledged "module" system in a future release.
Ideally, modules should encapsulate all primitives, reducing duplication
across the entire specification. The current implementation of
interactors is exclusive to the new interaction primitives and
generalizing it may necessitate changing/adding syntax. Thus, it is
better left to a future release.