Cleanup 2 - removed trailing spaces, last line new lines
This patch should not change any functionality,
but it ensures that there will be fewer changes
in the future as many editors do this automatically.

This way if someone commits a change, and they use
an editor (like IDEA or Atom or Sublime) that
automatically remove trailing spaces, people
will concentrate just on the change of code,
and won't be bothered with the changes in spaces.