Fix bound.js: Test if subgroup has bounds to union
In one of my Vega graphs, I have nested `group` marks. My data is filterable via signals. If too much data gets filtered, some subgroups may lose all their backing data and they lose the `bounds` property. Vega then crashes downstream of the `bounds.union()` call that I patched in this commit, when it tries to update the bounds of the parent group.

This may also have an effect on vega/vega#559 which seems to crash for the same reason.