Update marks.md
I could not find a case where the enter block is re-evaluated when the visualization is resized. 
For example, in a visualization I created that resizes based on the container width, the following encoding rule only takes effect when the marks are first instantiated. (The stroke changes color as expected when this style rule is placed in an update block.) 
{
  "enter": {
      "stroke":{"signal":"width < 400 ? 'red' : 'blue'"}
    }
}