Remove always null key from getExpression
As far as I can tell, there is never a `$name` param set on an expression parameter in the dataflow. However, Vega currently uses that property to compute a name for that expression. This removes that lookup and should be a no-op.

I am not sure why the `$name` property is there, I am guessing it may be a relic of an older expression implementation that had a name.