docs: Add 4 new examples (#3851)
**New Interaction Techniques examples**

Example 1: A simple one-column table with a vertical scrollbar. The
scrollbar responds to dragging the scrollbar with a mouse, mouse wheel
scrolling, and the Page Up, Page Down, Arrow Up, and Arrow Down buttons.

Example 2: Automatically switches between a column chart and a line
chart depending on the number of data points.

Example 3: An example of a custom tooltip. A map featuring a circular,
semi-transparent tooltip that displays information about the county
under the cursor.

**New Custom Design examples**

Example 4: A globe visualization of earthquakes. Enhanced and animated
version.

_Example 5: [DELETED FROM THIS PULL REQUEST] Table (Matrix): An example
of a table (matrix) with embedded images (country flags) and bar
charts._

**Other changes**

Warming Stripes example - data file updated (just for consistency with
the one in Vega datasets)

---------

Co-authored-by: Dominik Moritz <domoritz@gmail.com>