feat(vega-typings): add Typescript Types for `vega-loader` (#4000)
## Motivation

- I was learning about `vega-loader`, and decided to move the
information from the
[README](https://github.com/vega/vega/blob/main/packages/vega-loader/README.md)
to the typings file.
- This removes 3 `any` types from `Loader` too!

## Changes

- This should have no runtime impact, it only updates Typescript
definitions in `vega-typings`
- Note: `vega-loader`'s README does not mention that `sanitize` takes in
`context` and `response` as additional options - this is something I
noticed from `console.log`'ing the output of the options. For now I only
document that in the Typescript types.
- I could update the README for vega-loader as well if we want consumers
to know about this, or leave it out if this is supposed to be an
internal implementation detail.